requirements.txt
requests
beautifulsoup4
transformers==4.29.2
torch
faiss-cpu
numpy
scikit-learn
